base.php, call(), if handler not valid, issue fatal error, not 404
technically, the route was matched, but the handler was not found. therefore a 404 error is confusing. 

developer will be notified by a 500 error, specifying the class or method / function that was not callable

appears I could not use the error_get_last() to populate the $error array, since a real error was not fired, so I set that argument to NULL. Not sure if that has ramifications to debugging stack traces...