Preview/Template/View: added the `$this->f3` property discussed in #50 This patch adds the `View->f3` property discussed in https://github.com/bcosca/fatfree-core/pull/50. It passes the `View` and `Template` unit tests and is a better alternative than escaping and cloning the whole `Base` object (repeatedly) by defining `Base::instance()->set('f3', Base::instance());` or `$customHive = ['f3' => Base::instance(), /* … */]`. Also it's compatible with `PHP 5.3`, which [isn't able to escape and clone objects](http://fatfreeframework.com/base#recursive). Please keep in mind that only accessible non-static properties are escaped with `PHP >= 5.4` and that `Base->$hive` is a private property.
As we aren't interested in escaping and cloning the `Base` instance to use its methods, we shouldn't escape and clone it.

I'm not satisfied by this patch and would recommend to use `Base::instance()` instead of `$this->f3`, but it could be useful for others:

* https://github.com/bcosca/fatfree-core/pull/50#issuecomment-94781039
* https://github.com/bcosca/fatfree-core/pull/50#issuecomment-94674124

Now it's your turn to decide the fate of this pull request. Have fun!