Method uppercase to avoid route failure
My use case is with simpleid/simpleid project. When trying to log in with OpenID 2 on http://osm.org the 3rd party send a long POST with a `_method=post` in it. I totally don't know if it comes from OpenID standard or not. Nonetheless, this make the authentication failed, because of this parsing.

Is there any rationale behind considering this parameter as the real method?