Check active transaction before rollback/commit
As PHP 8 reports the actual transaction state in the MySQL PDO driver[1], running multiple statements with DB\SQL::exec() can break if any of those statements causes the transaction to end early (as is the case in things like schema changes). This was silently ignored in PHP 5/7, but under PHP 8 it can break without these additional checks.

[1]: https://www.php.net/manual/en/migration80.incompatible.php#migration80.incompatible.pdo-mysql